Unselect any previous selections
Use an event to trigger unselection of all cells before marking new selection.
This is a Sapper thing but window can be undefined if the code is run on the server end so add a check to ensure the event is only triggered on the client end.
I haven't actually seen what kind of error this may cause but not hard to understand so left in.